Moves bytecode to Java 1.6 for agent instrumentation
Java 1.6 is widely used and the agent needs to compile against 1.6 to be
able to instrument 1.6 runtime code.

While its possible to use higher source level to compile against lower
bytecode level, this is not really working correctly. This will be
improved in Java 9. until then it is kind of best practice to keep your
source and target level down. Also using animalsniffer on that level to
find usage of newer jdk classes (will also be fixed in 9)

This issue has been repeatedly raised in Brave, most recently from
@CodingFabian and @ivansenic