Renames Reporter to Sender, which accepts a list of encoded spans (#6)
Sender is a component which puts encoded spans onto the transport. As
discussed, there are use cases where other libraries or systems
may be in charge of marshalling spans.

The Sender interface includes the ability to do a health check. This is
necessary for fail-fast scenarios, such as when Kafka is down.

The Reporter interface will return in a future pull request.