Removes the B generic parameter from Encoder and AsyncReporter
Before, we had 3 generic type parameters:
* S - span model type
* B - buffer that ^^ encodes into, usually byte[]
* M - message type that a list of spans are encoded into, usually byte[]

The second parameter, B, was there to support more performant encoding,
for example, using pooled buffers. The theory was that since our json
encoder was slow, people might want to write their own.

As of zipkin 1.8, encoding speed of byte[] is exceptional, hundreds of
nanoseconds for a local span to a few microseconds for a complicated
json one. There is no longer a strong performance reason to encode spans
into something besides a byte[].

By removing the `B` generic parameter, we are left with `S` and `M`. The
message type is still valuable. For example, you can stream writes
instead of concatenating byte arrays, and of course Scribe requires us
to split spans into log-lines. `S` is needed due to other libraries
having different forms of the Span type, for example, thrift generated
classes.

Bottom-line is that we get a cognitive win by removing `B`, yet still
have flexibility in how to form messages.