Updates ReporterMetrics with stats used by Finagle and Jaeger
This adds `ReporterMetrics.updateQueuedSpans,updateQueuedBytes` and adds
the cause for `ReporterMetrics.incrementDroppedMessages`. These allow
AsyncReporter to produce metrics used by Finagle and Jaeger's tracers.

This is partially paid for by removing the metrics scope functionality.
Especially a client library shouldn't have a concern like this (even if
we do on the server).