Sends whatever we have on AsyncReporter.close (#46)
Before, we tried to fill up a message before returning from close. This
could end up blocking up to `messageTimeout`, especially if there aren't
many spans coming in.

Now, we send whatever we have. This also reduces the blocking time on
close to bounded at 1second vs whatever messageTimeout was. This is
because the latency is now related to how long it takes for a sender to
complete: we have no way of knowing this, so arbitrarily choose 1
second for `closeTimeout`.