Lightens the AwaitableCallback a bit
@Logic-32 and @anuraaga schooled me that when using Java concurrency
utilities in `java.util.concurrent`, you also inherit the thread
visibility properties ensured with synchronized and volatile symbols.

https://docs.oracle.com/javase/6/docs/api/java/util/concurrent/package-summary.html#MemoryVisibility

Given this, a volatile marker on a field whose state is dependent on
`CountDownLatch` should be redundant. This removes that volatile
qualifier accordingly.

cc other gear-heads @raphw @nicmunroe and @felixbarny for fun and good
measure!