Updates all test dependencies including CI and removes Retrolambda (#214)
This updates all aspects of the build including GitHub Actions versions, maven versions and test dependency versions. Most notably, this allows use of Java 11-17. Main versions are left for a different change.

To achieve this, we replace parts of the build that used RetroLambda to use maven profiles (similar to zipkin), as RetroLambda is no longer maintained. We also have to keep Java 11 in releases as later versions cannot compiler Java 1.6 bytecode anymore.

Signed-off-by: Adrian Cole <adrian@tetrate.io>