updates integration tests and benchmarks to use log4j
Similar to Brave, we don't want a large amount of clutter. To reduce a
lot of it, I moved anything that uses MockWebServer to an IT (because it
spams logs). I didn't move normal tests to log4j because we have a log
capturing test that would fail. If someone wants to take a pass later at
that, they could.

Signed-off-by: Adrian Cole <adrian@tetrate.io>