Deprecates Sender for much simpler BytesMessageSender
On yesterday's call with @shakuzen @anuraaga and @kojilin, we chatted
about how the async reporters never actually use the async code of the
senders. The async side was only used by zipkin-server, as the async
reporter design intentionally uses a blocking loop to flush the span
backlog. So, this makes things simpler by only requiring a blocking
implementation: `BytesMessageSender.send` (chosen to not create symbol
collisions and similar to `BytesMessageEncoder`.

The side effect is that new senders can use this interface and
completely avoid the complicated `Call` then `execute` chain, deferring
to whatever the library-specific blocking path is. I have implemented
this in armeria as a test, and it is absolutely easier.

As older spring libraries like sleuth may not upgrade, this maintains
the old types until reporter 4. However, new senders can simply
implement the new type and be done with it.

Signed-off-by: Adrian Cole <adrian@tetrate.io>