Adds SECURITY.md and scanning workflow
This adds SECURITY.md and a scanning workflow, using Trivy. In
particular, this clarifies what we use to scan for vulnerabilities
(Trivy, not anything else), and the only channel likely to be responded
to on a significant issue (zipkin-admin email, not advisories as people
ignored them).

I won't merge this until I get at least 2 approves. After that, I'll
help apply this to the other java repos.

Signed-off-by: Adrian Cole <adrian@tetrate.io>